Enemy Spawner
+ Add Enemy Spawner with the following functionality:
(1) Spawns enemies in a random position outside of the camera bounds.
(2) Keeps track of all Enemies that are currently spawned (hint: spawn them as children of the SpawnerObject and use GetChildCount to count how many are active).
(3) Stops spawning enemies when a max number of enemies parameter has been reached.
(4) Sends the OnShowChoiceUI event if the max number of enemies have been spawned and it’s child count is 0.
(5) Has a configurable “Time Between Spawns” float parameter.